+++ packages/Validation/src/Rules/BaseValidationRule.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Aedart\Validation\Rules;
+
+use Illuminate\Contracts\Validation\ValidationRule;
+
+/**
+ * Base Validation Rule
+ *
+ * @see \Illuminate\Contracts\Validation\ValidationRule
+ *
+ * @author Alin Eugen Deac <aedart@gmail.com>
+ * @package Aedart\Validation\Rules
+ */
+abstract class BaseValidationRule implements ValidationRule
+{
+}