+++ packages/Validation/src/Rules/Concerns/Translations.php
@@ -6,6 +6,8 @@ use Aedart\Support\Helpers\Translation\TranslatorTrait;
 use Aedart\Utils\Str;
 
 /**
+ * @deprecated Since 7.4 - Will be removed in next major version
+ *
  * Concerns Translations
  *
  * @author Alin Eugen Deac <aedart@gmail.com>